memcload/requirements.txt
protobuf==3.10.0
python-memcached==1.59
six==1.12.0
